Extract HasArguments
